+++ migrations/20170331135051-create-quiz-2.js
@@ -33,6 +33,9 @@ module.exports = {
           // onUpdate: 'cascade',
           // onDelete: 'cascade'
         }
+      },
+      {
+        freezeTableName: true
       }
     )
   },