+++ test/unit/vote.test.js
@@ -37,6 +37,33 @@ describe(title, () => {
     }
     query.vote(testVote).spread((result, created) => {
       assert.isTrue(created, 'vote should be created')
+      // console.log(result.dataValues)
+      done()
+    })
+  })
+
+  it('should be able to update the previous vote', (done) => {
+    let testVote = {
+      userId: 1,
+      quizId: 2,
+      stars: 0
+    }
+    query.vote(testVote).spread((result, created) => {
+      assert.isFalse(created, 'vote should not have been created')
+      // console.log(result)
+      done()
+    })
+  })
+
+  it('should not be able to update vote with bad params', (done) => {
+    let badVote = {
+      userId: 1,
+      quizId: 2,
+      error: 'duhhh'
+    }
+    query.vote(badVote).spread((result, created) => {
+      expect(result).to.have.any.keys('error')
+      assert.isFalse(created, 'vote should not have been created')
       done()
     })
   })